Add files via upload
4. String Handling using Library Functions:
a. Palindrome Check for a String: Write a program to check if a string is a palindrome or not.
b. String Length Calculation Using Functions: Write a program to calculate the length of a string using a library functions / predefined operators.
c. String Concatenation: Write a program to concatenate two strings using library functions.